+++ pykwalify/core.py
@@ -42,6 +42,7 @@ class Core(object):
         self.source = None
         self.schema = None
         self.validation_errors = None
+        self.validation_errors_exceptions = None
         self.root_rule = None
         self.extensions = extensions
 
@@ -141,7 +142,8 @@ class Core(object):
         log.debug("starting core")
 
         errors = self._start_validate(self.source)
-        self.validation_errors = errors
+        self.validation_errors = [str(error) for error in errors]
+        self.validation_errors_exceptions = errors
 
         if errors is None or len(errors) == 0:
             log.info("validation.valid")
@@ -150,7 +152,8 @@ class Core(object):
             log.error(" --- All found errors ---")
             log.error(errors)
             if raise_exception:
-                raise SchemaError("validation.invalid : {}".format(errors))
+                raise SchemaError("Schema validation failed:\n - {error_msg}.".format(
+                    error_msg='.\n - '.join(self.validation_errors)))
             else:
                 log.error("Errors found but will not raise exception...")
 
@@ -242,16 +245,21 @@ class Core(object):
     def _validate_include(self, value, rule, path, errors, done=None):
         # TODO: It is difficult to get a good test case to trigger this if case
         if rule._include_name is None:
-            errors.append("Include name not valid : {} : {}".format(path, value))
+            errors.append(SchemaError.SchemaErrorEntry(
+                msg='Include name not valid',
+                path=path,
+                value=value))
             return
 
         include_name = rule._include_name
         partial_schema_rule = pykwalify.partial_schemas.get(include_name, None)
         if not partial_schema_rule:
-            errors.append("No partial schema found for name : {} : Existing partial schemas: {}".format(
-                include_name,
-                ", ".join(sorted(pykwalify.partial_schemas.keys()))),
-            )
+            errors.append(SchemaError.SchemaErrorEntry(
+                msg="Cannot find partial schema with name '{include_name}'. Existing partial schemas: '{existing_schemas}'. Path: '{path}'",
+                path=path,
+                value=value,
+                include_name=include_name,
+                existing_schemas=", ".join(sorted(pykwalify.partial_schemas.keys()))))
             return
 
         self._validate(value, partial_schema_rule, path, errors, done)
@@ -281,8 +289,8 @@ class Core(object):
         ok_values = []
         error_tracker = []
 
-        unique_errors = set([])
-        map_unique_errors = set([])
+        unique_errors = {}
+        map_unique_errors = {}
 
         for i, item in enumerate(value):
             processed = []
@@ -324,7 +332,14 @@ class Core(object):
                                 if val in table:
                                     curr_path = "{}/{}/{}".format(path, j, v)
                                     prev_path = "{}/{}/{}".format(path, table[val], v)
-                                    map_unique_errors.add("value.notunique :: value: {} : {} : {}".format(val, curr_path, prev_path))
+                                    s = SchemaError.SchemaErrorEntry(
+                                        msg="Value '{duplicate}' is not unique. Previous path: '{prev_path}'. Path: '{path}'",
+                                        path=curr_path,
+                                        value=value,
+                                        duplicate=val,
+                                        prev_path=prev_path,
+                                    )
+                                    map_unique_errors[s.__repr__()] = s
                                 else:
                                     table[val] = j
                 elif r._unique:
@@ -338,7 +353,14 @@ class Core(object):
                         if val in table:
                             curr_path = "{}/{}".format(path, j)
                             prev_path = "{}/{}".format(path, table[val])
-                            unique_errors.add("value.notunique :: value: {} : {} : {}".format(val, curr_path, prev_path))
+                            s = SchemaError.SchemaErrorEntry(
+                                msg="Value '{duplicate}' is not unique. Previous path: '{prev_path}'. Path: '{path}'",
+                                path=curr_path,
+                                value=value,
+                                duplicate=val,
+                                prev_path=prev_path,
+                            )
+                            unique_errors[s.__repr__()] = s
                         else:
                             table[val] = j
 
@@ -433,7 +455,11 @@ class Core(object):
 
         for k, rr in m.items():
             if rr._required and k not in value:
-                errors.append("required.nokey : {} : {}".format(k, path))
+                errors.append(SchemaError.SchemaErrorEntry(
+                    msg="Cannot find required key '{key}'. Path: '{path}'",
+                    path=path,
+                    value=value,
+                    key=k))
             if k not in value and rr._default is not None:
                 value[k] = rr._default
 
@@ -463,18 +489,32 @@ class Core(object):
                         log.debug("Matched atleast one regex")
                     else:
                         log.debug("No regex matched")
-                        errors.append("key.regex.nomatch.any : {} : {}".format(k, "  ".join([mm[0]._map_regex_rule for mm in regex_mappings])))
+                        errors.append(SchemaError.SchemaErrorEntry(
+                            msg="Key '{key}' does not match any regex '{regex}'. Path: '{path}'",
+                            path=path,
+                            value=value,
+                            key=k,
+                            regex="  ".join([mm[0]._map_regex_rule for mm in regex_mappings])))
                 elif rule._matching_rule == "all":
                     if all(sub_regex_result):
                         log.debug("Matched all regex rules")
                     else:
                         log.debug("Did not match all regex rules")
-                        errors.append("key.regex.nomatch.all : {} : {}".format(k, "  ".join([mm[0]._map_regex_rule for mm in regex_mappings])))
+                        errors.append(SchemaError.SchemaErrorEntry(
+                            msg="Key '{key}' does not match all regex '{regex}'. Path: '{path}'",
+                            path=path,
+                            value=value,
+                            key=k,
+                            regex="  ".join([mm[0]._map_regex_rule for mm in regex_mappings])))
                 else:
                     log.debug("No mapping rule defined")
             elif r is None:
                 if not rule._allowempty_map:
-                    errors.append("key.undefined : {} : {}".format(k, path))
+                    errors.append(SchemaError.SchemaErrorEntry(
+                        msg="Key '{key}' was not defined. Path: '{path}'",
+                        path=path,
+                        value=value,
+                        key=k))
             else:
                 if not r._schema:
                     # validate recursively
@@ -495,7 +535,10 @@ class Core(object):
 
         if rule._enum is not None:
             if value not in rule._enum:
-                errors.append("enum.notexists : {} : {}".format(value, path))
+                errors.append(SchemaError.SchemaErrorEntry(
+                    msg="Enum '{value}' does not exist. Path: '{path}'",
+                    path=path,
+                    value=value))
 
         # Set default value
         if rule._default and value is None:
@@ -509,7 +552,11 @@ class Core(object):
         if rule._pattern is not None:
             res = re.match(rule._pattern, str(value))
             if res is None:  # Not matching
-                errors.append("pattern.unmatch : {} --> {} : {}".format(rule._pattern, value, path))
+                errors.append(SchemaError.SchemaErrorEntry(
+                    msg="Value '{value}' does not match pattern '{pattern}'. Path: '{path}'",
+                    path=path,
+                    value=value,
+                    pattern=rule._pattern))
 
         if rule._range is not None:
             if not is_scalar(value):
@@ -567,19 +614,39 @@ class Core(object):
 
         if max_ is not None:
             if max_ < value:
-                errors.append("{}.range.toolarge : {} < {} : {}".format(prefix, max_, value, path))
+                errors.append(SchemaError.SchemaErrorEntry(
+                    msg="Type '{prefix}' has size of '{value}', greater than max limit '{max_}'. Path: '{path}'",
+                    path=path,
+                    value=value,
+                    prefix=prefix,
+                    max_=max_))
 
         if min_ is not None:
             if min_ > value:
-                errors.append("{}.range.toosmall : {} > {} : {}".format(prefix, min_, value, path))
+                errors.append(SchemaError.SchemaErrorEntry(
+                    msg="Type '{prefix}' has size of '{value}', less than min limit '{min_}'. Path: '{path}'",
+                    path=path,
+                    value=value,
+                    prefix=prefix,
+                    min_=min_))
 
         if max_ex is not None:
             if max_ex <= value:
-                errors.append("{}.range.tolarge-ex : {} <= {} : {}".format(prefix, max_ex, value, path))
+                errors.append(SchemaError.SchemaErrorEntry(
+                    msg="Type '{prefix}' has size of '{value}', greater than or equals to max limit(exclusive) '{max_ex}'. Path: '{path}'",
+                    path=path,
+                    value=value,
+                    prefix=prefix,
+                    max_ex=max_ex))
 
         if min_ex is not None:
             if min_ex >= value:
-                errors.append("{}.range.toosmall-ex : {} >= {} : {}".format(prefix, min_ex, value, path))
+                errors.append(SchemaError.SchemaErrorEntry(
+                    msg="Type '{prefix}' has size of '{value}', less than or equals to min limit(exclusive) '{min_ex}'. Path: '{path}'",
+                    path=path,
+                    value=value,
+                    prefix=prefix,
+                    min_ex=min_ex))
 
     def _validate_scalar_type(self, value, t, errors, path):
         log.debug("Core scalar: validating scalar type : {}".format(t))
@@ -587,7 +654,11 @@ class Core(object):
 
         try:
             if not tt[t](value):
-                errors.append("Value: {} is not of type '{}' : {}".format(value, t, path))
+                errors.append(SchemaError.SchemaErrorEntry(
+                    msg="Value '{value}' is not of type '{scalar_type}'. Path: '{path}'",
+                    path=path,
+                    value=value,
+                    scalar_type=t))
         except Exception:
             # Type not found in map
             raise Exception("Unknown type check: {} : {} : {}".format(path, value, t))